requirements.txt
# Google API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# LLM & embeddings
langchain>=1.0.0
langchain-text-splitters
langchain-community
sentence-transformers
langchain-huggingface
langchain-ollama

# Vector DB
supabase

# Utils
packaging
python-dotenv

# LangGraph
langgraph


